lopif: be more explicit about 16-bit types
The following arrays are assumed to contain 16-bit entries,
so define them as "unsigned short" explicitly:
  xg_hival
  xg_and16
  xg_or16
  xg_s1
  xg_s2

The following function signatures were inconsistent,
so change the return values / arguments to match their use:
  xh_RgbHiPal
  xh_ByteToHi

+++ clemkgi.c
@@ -708,7 +708,7 @@ void z_bitbyte(unsigned char *buf1, unsigned char *buf2, int delka)
 //      palette of picture
 
 void xh_ByteToHi(unsigned char *Ibuf, unsigned char *Hi,
-		short Pixs, short Rows, short LenLine)
+		int Pixs, int Rows, int LenLine)
 {
 // Ibuf- (in) buffer with rows of palette picture (1B/pixel)
 // Hi  - (out)buffer with rows HiCol
